lyt 조금 수정

+++ ZOOTOPIA/src/main/java/com/aloha/zootopia/domain/InsuranceQna.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class InsuranceQna {
     private int qnaId;             // 질문 ID
     private int productId;         // 보험상품 ID
-    private int userId;            // 질문 작성자 ID
+    private long userId;            // 질문 작성자 ID
     private String species;        // 종류
     private String question;       // 질문 내용
     private String answer;         // 관리자 답변
